mysite/requirements.txt
absl-py==0.1.11
astor==0.6.2
bleach==1.5.0
decorator==4.2.1
Django==1.11
gast==0.2.0
grpcio==1.10.0
gunicorn==19.7.1
h5py==2.7.1
html5lib==0.9999999
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
Keras==2.1.5
Markdown==2.6.11
numpy==1.14.2
opencv-python==3.4.0.12
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2018.3
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==3.3.1
dj-database-url==0.5.0
psycopg2==2.7.1
